additional helmrepo and appsub content removal
