--- pkg/subscription/subscription_suite_test.go
@@ -1,13 +1,0 @@
-package subscription
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestSubscription(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Subscription Suite")
-}
